+++ warp/_src/fem/space/grid_3d_function_space.py
@@ -22,6 +22,7 @@ from warp._src.fem.polynomial import is_closed
 from warp._src.fem.types import ElementIndex
 
 from .shape import (
+    CubeBSplineShapeFunctions,
     CubeShapeFunction,
     CubeTripolynomialShapeFunctions,
 )
@@ -219,9 +220,70 @@ class GridTripolynomialSpaceTopology(SpaceTopology):
         return np.meshgrid(X, Y, Z, indexing="ij")
 
 
+class GridBSplineSpaceTopology(SpaceTopology):
+    def __init__(self, grid: Grid3D, shape: CubeBSplineShapeFunctions):
+        super().__init__(grid, shape.NODES_PER_ELEMENT)
+        self._shape = shape
+
+        self._padding = self._shape.PADDING
+
+        self.element_node_index = self._make_element_node_index()
+
+    def node_count(self) -> int:
+        return (
+            (self.geometry.res[0] + 2 * self._padding + 1)
+            * (self.geometry.res[1] + 2 * self._padding + 1)
+            * (self.geometry.res[2] + 2 * self._padding + 1)
+        )
+
+    def _make_element_node_index(self):
+        PADDING = self._padding
+
+        @cache.dynamic_func(suffix=self.name)
+        def element_node_index(
+            cell_arg: Grid3D.CellArg,
+            topo_arg: self.TopologyArg,
+            element_index: ElementIndex,
+            node_index_in_elt: int,
+        ):
+            res = cell_arg.res
+            cell = Grid3D.get_cell(res, element_index)
+
+            node_i, node_j, node_k = self._shape._node_ijk(node_index_in_elt)
+
+            node_x = cell[0] + node_i + PADDING
+            node_y = cell[1] + node_j + PADDING
+            node_z = cell[2] + node_k + PADDING
+
+            node_pitch_y = res[2] + 2 * PADDING + 1
+            node_pitch_x = node_pitch_y * (res[1] + 2 * PADDING + 1)
+            node_index = node_pitch_x * node_x + node_pitch_y * node_y + node_z
+
+            return node_index
+
+        return element_node_index
+
+    def node_grid(self):
+        res = np.array(self.geometry.res) + 2 * self._padding + 1
+
+        grid_coords_x = np.arange(0, res[0], dtype=float) - self._padding
+        X = grid_coords_x * self.geometry.cell_size[0] + self.geometry.origin[0]
+
+        grid_coords_y = np.arange(0, res[1], dtype=float) - self._padding
+        Y = grid_coords_y * self.geometry.cell_size[1] + self.geometry.origin[1]
+
+        grid_coords_z = np.arange(0, res[2], dtype=float) - self._padding
+        Z = grid_coords_z * self.geometry.cell_size[2] + self.geometry.origin[2]
+
+        return np.meshgrid(X, Y, Z, indexing="ij")
+
+
 def make_grid_3d_space_topology(grid: Grid3D, shape: CubeShapeFunction):
     if isinstance(shape, CubeTripolynomialShapeFunctions) and is_closed(shape.family):
         return forward_base_topology(GridTripolynomialSpaceTopology, grid, shape)
+
+    if isinstance(shape, CubeBSplineShapeFunctions):
+        return forward_base_topology(GridBSplineSpaceTopology, grid, shape)
 
     if isinstance(shape, CubeShapeFunction):
         return forward_base_topology(Grid3DSpaceTopology, grid, shape)